Chapter 12: Introduced Bank methapor in simple addition test
Test errors

Since Expression is a class used internally we don't get to show it in
the test. Since Ruby is not typed we cannot show that the sum temp var
is going to be an Expression.

Having a variable of type Expression declared in the test
helps envisioning our, not thoroughly thought, design.

As a mean towards our methapor (Bank + Expression) we could assert that
sum class is Expression. I decided not to do that, just because of
personal preference.