Chapter 13: Expression module as common interface
This is unnecessary in Ruby. However, it is a practice some people
follow in Ruby, to express what we would express with interfaces
in Java.

If we wanted to ensure certain classes implement an interface-like
module, then we would need explicit tests.

Read more about this topic in Sandi Metz' book:
  Practical Object-Oriented Design in Ruby: An Agile Primer