Chapter 14: Rates table in Bank
Test errors

Array in Ruby checks that elements are equal. Still used Pair to better
follow up with the examples.

There are no private classes in Ruby. They are usually represented
by nesting them. It is not always obvious they are not used outside
the scope they are declared in.

Faking hash returning zero would not work, I implemented it as it is
common in Ruby, calculating the hash of an Array containing all values.